+++ pages/_campus.vue
@@ -4,11 +4,12 @@
 
 <script>
 import { mapGetters } from 'vuex';
+import { DateTime } from 'luxon';
 
 export default {
   layout: ({ route: { name } }) => {
     let layout = 'default';
-    if (name === 'campus-rides' || name === 'campus-rides-planning') {
+    if (name === 'campus-rides' || name === 'campus-planificator') {
       layout = 'with-map';
     } else if (name === 'campus-map') {
       layout = 'simple';
@@ -20,7 +21,10 @@ export default {
   },
   async asyncData({ params, store }) {
     if (!store.getters['context/isCampus'](params.campus)) {
+      const start = DateTime.local().startOf('days').toJSDate();
+      const end = DateTime.local().endOf('days').toJSDate();
       await store.dispatch('context/fetchCampus', params.campus);
+      await store.dispatch('realtime/setRides', { campus: params.campus, start, end });
     }
     return {};
   },