+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * Add pins for POIs on the regulator map
 ### Fix
+* Add horizontal scroll in planificator when there is too much driver
 ## Version 1.4.2
 ### Features
 * Improve reliability, k8s side