+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ### Features
 * Add pins for POIs on the regulator map
 * Add logic of automatic static validation
+* Add planning vehicule and remove old system
 ### Fix
+* Add horizontal scroll in planificator when there is too much driver
 ## Version 1.4.2
 ### Features
 * Improve reliability, k8s side