+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix fetch ride when selecting a base
 * Fix tooltip overflow on planificator
 * Handle errors on post/patch /users /campuses/{id}/users /campuses/{id}/drivers
+* Fix planificator column alignment between header and body on month and day view
 ## Version 1.4.2
 ### Features
 * Improve reliability, k8s side